Tighten WorkflowState draft reducer
