+++ web/src/components/WorkflowState.tsx
@@ -19,7 +19,6 @@ import {
   updateCurrentState,
 } from "../util/api";
 import {
-  type ResolvedAdditionalField,
   getAdditionalFieldDefinitions,
 } from "../util/workflow";
 import { entryNameOrEmpty } from "../util/optionalValues";
@@ -27,12 +26,11 @@ import GlobalEntryField from "./GlobalEntryField";
 import AutosaveStatus from "./AutosaveStatus";
 import { useAutosave } from "./useAutosave";
 import { usePieceDetailSaveStatus } from "./usePieceDetailSaveStatus";
-
-export type ImageEntry = {
-  url: string;
-  caption: string;
-  cloudinary_public_id?: string | null;
-};
+import {
+  buildDraftState,
+  draftReducer,
+  normalizeAdditionalFieldPayload,
+} from "./workflowStateDraft";
 
 type WorkflowStateProps = {
   initialPieceState: PieceState;
@@ -42,167 +40,6 @@ type WorkflowStateProps = {
   autosaveDelayMs?: number;
 };
 
-type AdditionalFieldInputMap = Record<string, string>;
-type GlobalRefPkMap = Record<string, string>;
-type DraftState = {
-  baseState: PieceState;
-  notes: string;
-  images: ImageEntry[];
-  additionalFieldInputs: AdditionalFieldInputMap;
-  globalRefPks: GlobalRefPkMap;
-};
-type DraftAction =
-  | { type: "replace_base_state"; pieceState: PieceState }
-  | { type: "set_notes"; notes: string }
-  | { type: "set_additional_field"; name: string; value: string }
-  | { type: "set_global_ref_pks"; globalRefPks: GlobalRefPkMap };
-
-// Global-ref objects always carry an id and name. Any object without a name
-// is not a valid global ref value — treat it as empty so callers receive a
-// typed contract rather than a runtime fallback.
-function formatAdditionalFieldValue(
-  value: unknown,
-  type: ResolvedAdditionalField["type"],
-): string {
-  if (value === null || value === undefined) {
-    return "";
-  }
-  if (typeof value === "object") {
-    const obj = value as Record<string, unknown>;
-    if (typeof obj.name === "string") {
-      return obj.name;
-    }
-    // Objects without a name field are not representable as a string input.
-    return "";
-  }
-  if (type === "boolean") {
-    return value ? "true" : "false";
-  }
-  return String(value);
-}
-
-function extractGlobalRefPk(value: unknown): string | undefined {
-  if (typeof value === "object" && value !== null && "id" in value) {
-    const id = (value as { id: unknown }).id;
-    return typeof id === "string" ? id : undefined;
-  }
-  return undefined;
-}
-
-function buildAdditionalFieldInputMap(
-  defs: ResolvedAdditionalField[],
-  values: Record<string, unknown>,
-): AdditionalFieldInputMap {
-  const map: AdditionalFieldInputMap = {};
-  defs.forEach((def) => {
-    map[def.name] = formatAdditionalFieldValue(values[def.name], def.type);
-  });
-  return map;
-}
-
-function buildGlobalRefPkMap(
-  defs: ResolvedAdditionalField[],
-  values: Record<string, unknown>,
-): GlobalRefPkMap {
-  const map: GlobalRefPkMap = {};
-  defs.forEach((def) => {
-    if (def.isGlobalRef) {
-      const pk = extractGlobalRefPk(values[def.name]);
-      if (pk) map[def.name] = pk;
-    }
-  });
-  return map;
-}
-
-function normalizeAdditionalFieldPayload(
-  defs: ResolvedAdditionalField[],
-  inputs: AdditionalFieldInputMap,
-  globalRefPks: GlobalRefPkMap,
-): Record<string, string | number | boolean | null> {
-  const payload: Record<string, string | number | boolean | null> = {};
-  defs.forEach((def) => {
-    if (def.isGlobalRef) {
-      const pk = globalRefPks[def.name];
-      payload[def.name] = pk || null;
-      return;
-    }
-    // buildAdditionalFieldInputMap initializes all def names, so `raw` is
-    // always a string here. Guard is defensive for future call sites.
-    const raw = inputs[def.name] ?? "";
-    const trimmed = raw.trim();
-    if (trimmed === "") {
-      return;
-    }
-    if (def.type === "integer") {
-      const parsed = parseInt(trimmed, 10);
-      if (!Number.isNaN(parsed)) {
-        payload[def.name] = parsed;
-      }
-      return;
-    }
-    if (def.type === "number") {
-      const parsed = Number(trimmed);
-      if (!Number.isNaN(parsed)) {
-        payload[def.name] = parsed;
-      }
-      return;
-    }
-    if (def.type === "boolean") {
-      if (trimmed === "true") {
-        payload[def.name] = true;
-      } else if (trimmed === "false") {
-        payload[def.name] = false;
-      }
-      return;
-    }
-    payload[def.name] = raw;
-  });
-  return payload;
-}
-
-function stateImages(pieceState: PieceState): ImageEntry[] {
-  return pieceState.images.map((img) => ({
-    url: img.url,
-    caption: img.caption,
-    cloudinary_public_id: img.cloudinary_public_id ?? null,
-  }));
-}
-
-function buildDraftState(pieceState: PieceState): DraftState {
-  const additionalFieldDefs = getAdditionalFieldDefinitions(pieceState.state);
-  const additionalFields = pieceState.additional_fields ?? {};
-  return {
-    baseState: pieceState,
-    notes: pieceState.notes,
-    images: stateImages(pieceState),
-    additionalFieldInputs: buildAdditionalFieldInputMap(
-      additionalFieldDefs,
-      additionalFields,
-    ),
-    globalRefPks: buildGlobalRefPkMap(additionalFieldDefs, additionalFields),
-  };
-}
-
-function draftReducer(state: DraftState, action: DraftAction): DraftState {
-  switch (action.type) {
-    case "replace_base_state":
-      return buildDraftState(action.pieceState);
-    case "set_notes":
-      return { ...state, notes: action.notes };
-    case "set_additional_field":
-      return {
-        ...state,
-        additionalFieldInputs: {
-          ...state.additionalFieldInputs,
-          [action.name]: action.value,
-        },
-      };
-    case "set_global_ref_pks":
-      return { ...state, globalRefPks: action.globalRefPks };
-    default:
-      return state;
-  }
-}
 
 // ── ImageUploader ─────────────────────────────────────────────────────────────
 
@@ -316,6 +153,7 @@ export default function WorkflowState({
     buildDraftState,
   );
   const { baseState, notes, images, additionalFieldInputs, globalRefPks } = draft;
+  const baseDraft = useMemo(() => buildDraftState(baseState), [baseState]);
   const additionalFieldDefs = useMemo(
     () => getAdditionalFieldDefinitions(baseState.state),
     [baseState.state],
@@ -333,18 +171,10 @@ export default function WorkflowState({
     () =>
       normalizeAdditionalFieldPayload(
         additionalFieldDefs,
-        draft.baseState.additional_fields
-          ? buildAdditionalFieldInputMap(
-              additionalFieldDefs,
-              draft.baseState.additional_fields,
-            )
-          : {},
-        buildGlobalRefPkMap(
-          additionalFieldDefs,
-          draft.baseState.additional_fields ?? {},
-        ),
+        baseDraft.additionalFieldInputs,
+        baseDraft.globalRefPks,
       ),
-    [additionalFieldDefs, draft.baseState.additional_fields],
+    [additionalFieldDefs, baseDraft.additionalFieldInputs, baseDraft.globalRefPks],
   );
   const additionalFieldsDirty =
     JSON.stringify(normalizedAdditionalFields) !==
